+++ doc/scripts/build_par_rst.py
@@ -16,7 +16,7 @@ from pkg_resources import resource_filename
 from pypeit.par import pypeitpar
 from pypeit.par.parset import ParSet
 from pypeit.spectrographs.util import load_spectrograph
-from pypeit.spectrographs import valid_spectrographs
+from pypeit.defs import pypeit_spectrographs
 
 #-----------------------------------------------------------------------------
 #def class_name(p):
@@ -82,8 +82,7 @@ if __name__ == '__main__':
                            ':class:`pypeit.spectrographs.spectrograph.Spectrograph` class.', 72)
     lines += ['']
 
-    spectrographs = valid_spectrographs.pypeit_spectrographs
-    for spec in spectrographs:
+    for spec in pypeit_spectrographs:
         s = load_spectrograph(spec)
         lines += [ ' '.join([s.telescope['name'], s.camera, '(``{0}``)'.format(s.spectrograph)]) ]
         lines += [ '-'*len(lines[-1]) ]